+++ mongoengine/queryset/base.py
@@ -953,11 +953,12 @@ class BaseQuerySet:
             field = self._fields_to_dbfields([field]).pop()
         except LookUpError:
             pass
+
+        raw_values = queryset._cursor.distinct(field)
         if not self._auto_dereference:
-            return queryset._cursor.distinct(field)
-        distinct = self._dereference(
-            queryset._cursor.distinct(field), 1, name=field, instance=self._document
-        )
+            return raw_values
+
+        distinct = self._dereference(raw_values, 1, name=field, instance=self._document)
 
         doc_field = self._document._fields.get(field.split(".", 1)[0])
         instance = None